+++ changelog.md
@@ -8,6 +8,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ------------------
+## [2.1.0] - 2026-01-16
+### Added
+* Setting to select which flow to use for KEC when both one step checkout and two step flows are available and a AP key exists.
+
+### Fix
+* Fixed an issue where the one step checkout flow would not be prioritized when both flows were available.
+
 ## [2.0.0] - 2025-12-03
 ### Added
 * Implemented support for One step checkout flow for AP partners using Klarna Express Checkout.